fix(state): set `noticeCond` correctly in `ReadState` (#320)
With the addition of support for notices in the pebble state, there was a small bug introduced in `ReadState`, which was not updated to set the new `noticeCond`. As a result, a call to `WaitNotices` from a state which was created via `ReadState` will cause a null pointer exception.

This PR adds the proper `noticeCond` initialization in `ReadState`, as well as a unit test to check that `WaitNotices` does not panic on a state which was created via `ReadState`.