feat(state,daemon,client): add user-specific notices
We wish to be able to restrict the visibility of a given notice to
particular users. This PR adds to `Notice` a `user` field which
corresponds to the UID of the intended recipient of the notice. The root
user user has access to all notices, and is the only user which may view
notices with a `user` of 0. A `user` of -1 indicates that any user may
receive that notice.

Much like `types` and `keys`, requests to the notices API may include a
filter on `users`. Unlike the former, however, the API performs extra
checks to ensure that the filter which is passed into the state code
only contains UIDs for notices that the user which triggered the API
request is allowed to view. Notably, the state code is not aware of the
special handling of requests by particular users or notices with UID -1;
all the logic for checking the permissions of the request occur in the
notices API code. If a request comes in with a `users` filter including
UIDs of notices to which the requesting user does not have access, the
API prunes these from the filter, rather than throwing an error.

While non-root users may only receive notices with a `user` matching
their UID or -1, notices may be created with an arbitrary `user` as the
recipient (so long as that value is -1 or a valid uint32).

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>