feat(state,daemon,client): make notice user optional
Replace the mandatory `user int64` in notices with `userID *int` in
`AddNoticeOptions`, thus allowing the user ID to be omitted when
creating a notice. As a result, `user` is no longer a component of a
notice's `uniqueKey`, returning the behaviour more closely to that prior
to the addition of user-specific notices.

If a user ID is not specified (left `nil`) when creating a notice, then
the accessibility of that notice will correspond to the default notice
user ID, which for pebble is -1 (viewable by all).

When a new notice is added with a type and key matching that of an
existing notice, the user ID of the new notice must also match that of
the existing notice, or an error will be returned.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>