fix(daemon): handle invalid notice types in filter (#329)
Invalid notice types are ignored when parsing the API request to
construct the notice filter. However, if only invalid notice types are
included in the request, discarding all types results in a filter
indistinguishable from one which never had any types in the first place,
and thus should return notices of any type. This commit fixes this
behaviour, so no notices are returned if only invalid notice types are
requested.