feat(cli): implement personality for application name-agnostic help manuals (#238)
Implements the concept of personality, two variables that contain
information about the application identity, such as the program
binary name and the application display name. This enables the
implementation of Pebble-based software that does not necessarily
have the same name as Pebble.

Help strings throughout the application have been slightly modified
so self-references to `pebble`/Pebble are removed without changing
the semantics of the help manual.