feat: make shutdown action return meaningful exit code (#336)
This PR makes Pebble shutdowns due to `on-failure: shutdown` return exit code
10 instead of always returning exit code 0. The code 10 was chosen to keep
it "a few away" from other Pebble exit codes.

Similarly, if `on-check-failure` is used with `shutdown`, return exit code 11
instead of always returning code 0. (The 11 to distinguish a "check failure"
shutdown from a "service failure" shutdown.

It also adds support for `on-failure: success-shutdown` (also for
`on-check-failure`) and `on-success: failure-shutdown`, which reverses
the "polarity" of the exit code for those cases.

Both of these are as per spec [DA062]
(https://docs.google.com/document/d/1OYSjmL2_1Na9HTwYmYo4KWvlQC9fcusXdMPzpHRBP8k/edit).

Fixes #324.