feat(many): reimplement per-user notices using `userID *uint32`
This corresponds to the latest revision of the per-user notices spec,
and involves dramatic simplifications to the API, client, and cli. Now,
each notice has a user ID, stored as a `*uint32`, which must be
explicitly included when adding the notice to the state. If the user ID
is `nil`, then that notice is considered to be public. Non-admin users
only have access to their own notices and public notices, while admin
users may view any notice.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>