ci(snap): cope with concurrent builds
At the moment, the Pebble release process involves
a series of 3 commits under a very short amount of
time. GitHub's concurrency flag is being used to
cope with this, but it is unable to wait for
pending workflow runs, which raises the risk of
a Release workflow cancelling the previous
commit workflow, where the snap is being built,
if the run is pending, waiting for a GH runner.

This commit forces the snap build to happen on
releases also.