feat(planstate): create a dedicated plan manager
Move plan management out of the service manager.

The plan manager is responsible for appending and combining layers
as runtime changes are requested over the daemon API or internally
through tasks.

This is the first step towards a bigger goal of allowing the plan
to be extended from pebble derivative projects, as new external
managers add plan sections at startup time.

It is important to notice a simple pattern with managers receiving
plan updates: plan consumers (managers) should not mutate the plan
they get, as this is a local copy of the plan pointer, and cannot
affect the actual plan as maintained by the plan manager. This also
means the plan lock is only protecting local manager plan pointer
updates, and not needed for the content of the plan. The current
design of the plan manager logic continues to follow the previous
pattern of always creating a new plan instance on any update. This
means existing instances of plan is never mutated, and hence why
consumers can safely access plan snapshot content.