fix(enter): disable logs for interactive exec runs
This commit disables logs completely for ``pebble enter --verbose exec
-it <cmd..>`` even if the --verbose flag is passed.

Interactive applications, like Vim, can be run as an exec command:

    pebble enter --verbose exec -it vim

Since enter runs the pebble daemon and executes the exec command
afterwards, the same TTY that will be made RAW by exec to run
interactive applications like Vim is also used for logging and service
outputs. Thus, the logs coming from pebble find the TTY RAW and a weird
indentation is introduced. [1]

The issue however is not limited to only indentation. For example, if a
text editor was being run with exec -it, like Vim, the incoming logs
would go to the text editor and may change the content of the editor’s
buffer.

This commit disables logs completely for ``pebble enter --verbose exec
-it`` even if the --verbose flag is passed. Since logs are currently
disabled for ``pebble enter exec`` if --verbose flag is not passed and
-i, -t are assumed if stdin and stdout are TTYs, the commit only ignores
the --verbose flag in cases where exec will make the TTY RAW.

References:
- [1] https://github.com/canonical/pebble/issues/339