fix(daemon): improve health state lock test, remove LockCount
As Harry pointed out at
https://github.com/canonical/pebble/pull/369#discussion_r1505423350,
there's a much simpler way to test this without needing a new State
method like LockCount. Just acquire the state lock, then call the
endpoint. If it times out, we know it was trying to acquire the lock.

In addition, fix an issue where the health endpoint would still hold
the state lock if it returned an error. Fix those and add a test for
that too.