feat(cli): add RunOptions to cli.Run, remove global variable (#383)
The goal of this PR is to allow full customization of the
`client.Config` (including its `Socket`, `BaseURL` and `UserAgent`
members) for callers of `cli.Run()`. This is achieved by adding a
`RunOptions` struct as argument to `cli.Run()`.

The `clientConfig` global variable isn't accessed by any non-test code,
so it can be removed.

---

Additional information about test-related changes:

The global variable WAS used in tests (as `ClientConfig`, in
`internals/cli/cli_test.go`: `cli.ClientConfig.BaseURL = server.URL`),
but since all tests use `cli.RunMain()` (which is defined in
`cli/export_test.go`) instead of `cli.Run()`, we can just define the
global variable for the tests only, and make `RunMain` use that config
object by calling `Run` with the test `ClientConfig`.