feat(plan): disable KnownFields for YAML unmarshal
Disable YAML KnownFields during plan unmarshal, so that we align with
the general Pebble JSON unmarshal strategy (DisallowUnknownFields is false)

Pebble currently sets KnownFields=true when loading the configuration
layers, or when parsing a layer supplied over the daemon server API.

The result of this is strict attribute checking, which results in an
unmarshal error if any attribute present in the YAML does not exist
in the backing structure.

This option trades forward/backwards compatibility for user input
error detection. This may have been the best choice up until this point
in time.

However, Pebble is getting used in derivative projects where the ability
for the derivative binary to cope with an earlier or later plan revision
YAML data will make it more robust against upgrades and downgrades.

Since this relaxes the unmarshal rules, this will be backwards compatible
in all existing non-error use-cases. Any existing tests expecting an
error from Pebble will have to be updated/removed.

This would also align with how JSON decoding is done in Pebble (since we
never set DisallowUnknownFields).