feat(cli): new Info category for help, version (#400)
The commands `help` and `version` are not related to running the service
manager (you can run `help` also for client help, and `version` prints
both the client and the server version). This splits those two commands
into a new "Info" category, leaving only the `run` command in the "Run"
category.