fix(daemon): require admin access for file pull API
This locks down the files "pull" API to require admin. Even though it's
a read (GET), this meant any user could potentially read sensitive
files.

We also lock down the task-websocket endpoint to admin. This is a GET
endpoint, but these websockets are used by exec to send stdin/out/err
and commands to the exec'd process, so they should require admin too.