fix(daemon): require admin access for POSTs and file pull API
Most of this was introduced in PR #358, when we ported the
AccessChecker changes from snapd, but accidentally set all the
WriteAccess fields to UserAccess{} instead of AdminAccess{}.
Previously there was a r.Method=="GET" check in Command.canAccess that
handled this case.

Additionally:

- We lock down the files "pull" API to require admin. Even though it's
  a read (GET), this meant any user could potentially read sensitive
  files.
- We lock down the task-websocket endpoint to admin. This is a GET
  endpoint, but these websockets are used by exec to send stdin/out/err
  and commands to the exec'd process, so they should require admin too.

I've added some tests for these to ensure we don't accidentally change
them in future, without noticing. How valuable these tests are I'm not
sure, as they only cover a subset of the API endpoints, but it seems
better than nothing.