fix(checkstate): ignore and abort carryover tasks
Running checks have changes that gets persisted the state-engine.
This means that following a reboot, changes and tasks that are not
complete (ready) will be resumed. In the case of some managers, such
as checkstate, carryover changes and tasks are an unwanted side effect.

Currently the plan manager will perform the first plan load very early
during startup (before the state-engine is ready, and before StartUp
hooks have been called). The result is that a PlanChanged propagation
will take place at plan load, and force checkstate to inspect the
running checks prematurely.

Checkstate discovers a running change from a previous boot context,
and tries to load its data from cached state, which does not exist.

Gracefully ignore changes for which cached state does not exist,
and use this to identify changes that should be aborted on the first
ensure pass.