perf(daemon): avoid acquiring state lock (twice!) for every API request
In daemon.Command.ServeHTTP, at the end of every request we acquire the
state lock twice, once for including the maintenance (restart) info in
the response, and once for including the warnings summary. We don't use
either maintenance/restarts or warnings in Pebble at all, so let's just
turn that off.

We could rip out those features completely, but the APIs need to stay
for backwards compatibility, and it's possible we'll want them later,
so let's just disable them and their tests for now.

Fixes #366