fix: actually turn off child subreaping in reaper.Stop (#454)
Previously we weren't actually turning off child subreaping at the OS
level in `reaper.Stop`, so it didn't reverse `reaper.Start`. Fix this,
and also simplify the error handling a little.

This almost certainly didn't cause problems, because as soon as the
Pebble process exits it doesn't matter. But it wouldn't have turned it
off in tests, and it's just the Right Thing To Do.

Fixes #412