fix: send EOF to master pty in interactive mode
If the interactive mode is enforced, there are
scenarios like:

go run ./cmd/pebble exec -i -- tee < ./HACKING.md

that would prompt the server to shutdown the
execution earlier than the whole output from 'tee'
will be read and sent to the client. Hence,
instead of closing the master pty descriptor when
there is no more input from the client, send
Ctrl-D (EOF) to indicate to the process that there
will be no more input.