refactor(state): avoid need for state lock in LatestWarningTime
This means that not every request/response needs to acquire the state
lock, as shown by the removal of the custom response type in the
/v1/health endpoint.

Note that there's a test, TestHealthStateLockNotHeldSuccess, that
specifically tests that the state lock is not held during /v1/health
requests. If we accidentally introduce locking again, that will fail.