feat(daemon): introduce "external" reboot mode
This reboot mode ensures that the machine is not restarted during
daemon.Stop() when a restart has been requested. Instead, it signals the
daemon.Stop() caller that further actions to ensure a restart are
required through a new exported error daemon.ErrRestartExternal.

This is required in scenarios where further actions are required before
the machine can be safely rebooted, such as syncing and unmounting
filesystems.