fix: allow stopping services in "starting" state (#503)
Pebble has a 1-second okayWait time period for starting services, after
which, if the service is still not exited, it's considered as "running".
Previously, when services were still in the starting state within this
1-second okayWait period, they couldn't be stopped. This causes some
issues in tests. And, just because services have just started doesn't
mean they shouldn't be allowed to stop. This PR allows services in the
starting state (within the 1s okayWait period) to be stopped, no matter
if it's the daemon exits (Ctrl+C in the console) or the `pebble stop`
command is issued.